// types/branding.type.ts
export type Branding = {
  telephone_1: string;
  telephone_2: string;
  contact_email: string;
  website: string;
  organization_logo_url: string | null;
  organization_name: string;
  terms_conditions: string | null;
  theme: {
    primary: string;
    secondary: string;
    cardPrimary: string;
    backgroundPrimary: string;
  };
};